# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminLoginRequest:
    properties:
      email:
        example: admin@gmail.com
        type: string
      password:
        example: password123
        type: string
    type: object
  dto.AdminSignupRequest:
    properties:
      business_id:
        example: 1
        type: integer
      email:
        example: admin@gmail.com
        type: string
      name:
        example: Srajal
        type: string
      password:
        example: password123
        type: string
    type: object
  dto.CreateBookingRequest:
    properties:
      customer:
        description: Nested customer info for quick lookup or creation
        properties:
          name:
            example: Rahul
            type: string
          phone:
            example: "9876543210"
            type: string
        type: object
      notes:
        example: Haircut booking
        type: string
      service_id:
        example: 1
        type: integer
      staff_id:
        example: 2
        type: integer
      start_time:
        example: "2026-03-03T11:00:00Z"
        type: string
    type: object
  dto.CreateServiceRequest:
    properties:
      duration:
        description: Duration in minutes
        example: 30
        type: integer
      name:
        description: Service name
        example: Hair Cut
        type: string
      price:
        description: Service price
        example: 250
        type: number
    type: object
  dto.CreateStaffRequest:
    properties:
      name:
        description: Staff member full name
        example: Staff One
        type: string
      phone:
        description: Contact number (used for login)
        example: "9999999999"
        type: string
      pin:
        description: PIN for staff login (will be hashed)
        example: "1234"
        type: string
      role:
        description: Position/role in salon
        example: barber
        type: string
    type: object
  dto.SetAvailabilityRequest:
    properties:
      day_of_week:
        description: Day (0=Sunday, 1=Monday, ..., 6=Saturday)
        example: 1
        type: integer
      end_time:
        description: Work end time (HH:mm format)
        example: "18:00"
        type: string
      start_time:
        description: Work start time (HH:mm format)
        example: "09:00"
        type: string
    type: object
  dto.StaffLoginRequest:
    properties:
      phone:
        example: "9876543210"
        type: string
      pin:
        example: "1234"
        type: string
    type: object
  dto.UpdateServiceRequest:
    properties:
      duration:
        description: Updated duration in minutes
        example: 45
        type: integer
      name:
        description: Updated service name
        example: Hair Spa
        type: string
      price:
        description: Updated service price
        example: 600
        type: number
    type: object
host: localhost:3000
info:
  contact:
    email: srajal@example.com
    name: Srajal
  description: Multi-tenant Salon Management Backend built with Go + Fiber
  termsOfService: http://example.com/terms/
  title: Salon Management Backend API
  version: "1.0"
paths:
  /admin/bookings:
    post:
      consumes:
      - application/json
      description: Create a new booking (Admin only)
      parameters:
      - description: Booking Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - Bookings
  /admin/bookings/{id}/approve:
    put:
      description: Approve a pending booking (Admin only)
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Approve booking
      tags:
      - Bookings
  /admin/bookings/{id}/cancel:
    put:
      description: Cancel a booking (Admin only)
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel booking
      tags:
      - Bookings
  /admin/login:
    post:
      consumes:
      - application/json
      description: Login as admin
      parameters:
      - description: Admin Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Admin login
      tags:
      - Authentication
  /admin/services:
    post:
      consumes:
      - application/json
      description: Creates a new salon service for the authenticated business (Admin
        only)
      parameters:
      - description: Service Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new service
      tags:
      - Services
  /admin/services/{id}:
    delete:
      description: Soft deletes a service by marking it inactive (Admin only)
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete service
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Updates an existing service for the authenticated business (Admin
        only)
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update service
      tags:
      - Services
  /admin/signup:
    post:
      consumes:
      - application/json
      description: Register a new admin account
      parameters:
      - description: Admin Signup
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminSignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Admin signup
      tags:
      - Authentication
  /admin/staff:
    post:
      consumes:
      - application/json
      description: Create a new staff member (Admin only)
      parameters:
      - description: Staff Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create staff
      tags:
      - Staff
  /admin/staff/{staffId}/availability:
    post:
      consumes:
      - application/json
      description: Define working hours for a staff member
      parameters:
      - description: Staff ID
        in: path
        name: staffId
        required: true
        type: integer
      - description: Availability data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.SetAvailabilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Set staff availability
      tags:
      - Staff Availability
  /admin/staff/{staffId}/services/{serviceId}:
    delete:
      description: Remove an assigned service from a staff member (Admin only)
      parameters:
      - description: Staff ID
        in: path
        name: staffId
        required: true
        type: integer
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove service from staff
      tags:
      - Staff Services
    post:
      description: Assign a service to a staff member (Admin only)
      parameters:
      - description: Staff ID
        in: path
        name: staffId
        required: true
        type: integer
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign service to staff
      tags:
      - Staff Services
  /bookings:
    get:
      description: Get bookings (Admin or Staff)
      parameters:
      - description: Filter by date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: Filter by staff ID
        in: query
        name: staff_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List bookings
      tags:
      - Bookings
  /services:
    get:
      description: Returns all active services for the logged-in business
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List services
      tags:
      - Services
  /services/{id}:
    get:
      description: Returns a single active service belonging to the logged-in business
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get service by ID
      tags:
      - Services
  /staff/login:
    post:
      consumes:
      - application/json
      description: Login as staff and receive JWT token
      parameters:
      - description: Staff Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.StaffLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Staff login
      tags:
      - Authentication
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your_jwt_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"